pyproject.toml
[project]
name = "2dovmodel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "pillow>=11.2.1",
]
